+++ packages/js/src/plugins/rfqModule/operations/confirmResponse.ts
@@ -12,6 +12,7 @@ import {
   Signer,
   makeConfirmOptionsFinalizedOnMainnet,
 } from '../../../types';
+import { convertOverrideLegMultiplierBps } from '../helpers';
 import { TransactionBuilder, TransactionBuilderOptions } from '../../../utils';
 
 const Key = 'ConfirmResponseOperation' as const;
@@ -175,13 +176,14 @@ export const confirmResponseBuilder = async (
   options: TransactionBuilderOptions = {}
 ): Promise<TransactionBuilder> => {
   const { programs, payer = convergence.rpc().getDefaultFeePayer() } = options;
-  const {
-    taker = convergence.identity(),
-    rfq,
-    response,
-    side,
-    overrideLegMultiplierBps = null,
-  } = params;
+  const { taker = convergence.identity(), rfq, response, side } = params;
+  let { overrideLegMultiplierBps = null } = params;
+
+  if (overrideLegMultiplierBps) {
+    overrideLegMultiplierBps = convertOverrideLegMultiplierBps(
+      Number(overrideLegMultiplierBps)
+    );
+  }
 
   const responseModel = await convergence
     .rfqs()