+++ packages/js/src/plugins/rfqModule/operations/createAndFinalizeRfqConstruction.ts
@@ -5,9 +5,10 @@ import * as anchor from '@project-serum/anchor';
 import { SendAndConfirmTransactionResponse } from '../../rpcModule';
 import { SpotInstrument } from '../../spotInstrumentModule';
 import { OrderType, QuoteAsset, FixedSize } from '../types';
+import { assertRfq, Rfq } from '../models';
 import { PsyoptionsAmericanInstrument } from '../../psyoptionsAmericanInstrumentModule';
 import { PsyoptionsEuropeanInstrument } from '../../psyoptionsEuropeanInstrumentModule';
-import { Rfq } from '../models';
+
 import {
   instrumentsToLegsAndExpectedLegsHash,
   convertFixedSizeInput,
@@ -212,6 +213,7 @@ export const createAndFinalizeRfqConstructionOperationHandler: OperationHandler<
       const rfq = await convergence
         .rfqs()
         .findRfqByAddress({ address: rfqPda });
+      assertRfq(rfq);
 
       return { ...output, rfq };
     },