+++ packages/js/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @convergence-rfq/sdk
 
+## 4.0.24-rc.5
+
+### Patch Changes
+
+- use getAccount and createToken in prepareSettlement instead of getOrCreateAssociatedTokenAccount
+
+## 4.0.24-rc.4
+
+### Patch Changes
+
+- call getMultipleAccounts with only leq 100 pubkeys
+
+## 4.0.24-rc.3
+
+### Patch Changes
+
+- use getMultipleAccounts to fetch rfq, response, fix psyop euro fixed base rfq creation collateral
+
 ## 4.0.23-rc.6
 
 ### Patch Changes